feat: Exception handling - delete

+++ src/main/java/com/gusdev/spring_ecommerce_api/services/UserService.java
@@ -2,8 +2,10 @@ package com.gusdev.spring_ecommerce_api.services;
 
 import com.gusdev.spring_ecommerce_api.entities.User;
 import com.gusdev.spring_ecommerce_api.repositories.UserRepository;
+import com.gusdev.spring_ecommerce_api.services.exceptions.DatabaseException;
 import com.gusdev.spring_ecommerce_api.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -32,7 +34,15 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        userRepository.deleteById(id);
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
+
     }
 
     public User update(Long id, User obj) {
